Rename appendage to sequence
